+++ src/shared/components/index.ts
@@ -1,3 +1,4 @@
+export * from './AppInfo';
 export * from './AppSidebar';
 export * from './ButtonCounter';
 export * from './Markdown';