+++ src/showcase/examples/ButtonPage.tsx
@@ -1,34 +1,47 @@
 import React, { Fragment, useEffect } from 'react';
-import { IconButton, IIconButtonProps, homeSolidSvg, FormControl, COLOR, VARIANT } from '../../lib';
-import { IShowcaseBaseProps, withOptions } from './components';
+import { Button, COLOR, FormControl, IButtonProps, snackbarService, VARIANT } from '../../lib';
+import { withOptions } from './components';
+import { IShowcaseBaseProps } from './components';
 
-const ShowcaseIconButtonBase = ({ settingValues, setSettingsControls }: IShowcaseBaseProps<IIconButtonProps>) => {
+const ButtonPageBase = ({ settingValues, setSettingsControls }: IShowcaseBaseProps<IButtonProps>) => {
 
 	useEffect(() => {
 		setSettingsControls({
+			block: new FormControl(settingValues.block, [], 'checkbox', { label: 'block' }),
 			className: new FormControl(settingValues.className, [], 'text', { label: 'className', placeholder: 'css class' }),
 			color: new FormControl(settingValues.color, [], 'select', { label: 'color', options: Object.keys(COLOR).map(c => ({ label: c, value: c })) }),
 			disabled: new FormControl(settingValues.disabled, [], 'checkbox', { label: 'disabled' }),
 			isActive: new FormControl(settingValues.disabled, [], 'checkbox', { label: 'isActive' }),
+			isRounded: new FormControl(settingValues.disabled, [], 'checkbox', { label: 'isRounded' }),
 			variant: new FormControl(settingValues.variant, [], 'select', { label: 'variant', options: Object.keys(VARIANT).map(c => ({ label: c, value: c })) }),
 		});
 	}, []);
 
+	const handleClick = () => {
+		snackbarService.show('Button clicked');
+	}
+
 	return (
 		<Fragment>
-			<IconButton
-				className={settingValues.className}
+
+			<Button
+				block={settingValues.block}
 				color={settingValues.color}
-				isActive={settingValues.isActive}
+				className={settingValues.className}
 				disabled={settingValues.disabled}
-				icon={homeSolidSvg}
+				isActive={settingValues.isActive}
+				isRounded={settingValues.isRounded}
 				variant={settingValues.variant}
-			/>
+				onClick={handleClick}
+			>
+				some button text
+			</Button>
+
 		</Fragment>
 	);
 }
 
-export const ShowcaseIconButton = withOptions<IIconButtonProps>(ShowcaseIconButtonBase, {
+export const ButtonPage = withOptions<IButtonProps>(ButtonPageBase, {
 	color: COLOR.primary,
 	variant: VARIANT.normal
 });