+++ src/showcase/examples/IconButtonPage.tsx
@@ -1,47 +1,34 @@
 import React, { Fragment, useEffect } from 'react';
-import { Button, COLOR, FormControl, IButtonProps, snackbarService, VARIANT } from '../../lib';
-import { withOptions } from './components';
-import { IShowcaseBaseProps } from './components';
+import { IconButton, IIconButtonProps, homeSolidSvg, FormControl, COLOR, VARIANT } from '../../lib';
+import { IShowcaseBaseProps, withOptions } from './components';
 
-const ShowcaseButtonBase = ({ settingValues, setSettingsControls }: IShowcaseBaseProps<IButtonProps>) => {
+const IconButtonPageBase = ({ settingValues, setSettingsControls }: IShowcaseBaseProps<IIconButtonProps>) => {
 
 	useEffect(() => {
 		setSettingsControls({
-			block: new FormControl(settingValues.block, [], 'checkbox', { label: 'block' }),
 			className: new FormControl(settingValues.className, [], 'text', { label: 'className', placeholder: 'css class' }),
 			color: new FormControl(settingValues.color, [], 'select', { label: 'color', options: Object.keys(COLOR).map(c => ({ label: c, value: c })) }),
 			disabled: new FormControl(settingValues.disabled, [], 'checkbox', { label: 'disabled' }),
 			isActive: new FormControl(settingValues.disabled, [], 'checkbox', { label: 'isActive' }),
-			isRounded: new FormControl(settingValues.disabled, [], 'checkbox', { label: 'isRounded' }),
 			variant: new FormControl(settingValues.variant, [], 'select', { label: 'variant', options: Object.keys(VARIANT).map(c => ({ label: c, value: c })) }),
 		});
 	}, []);
 
-	const handleClick = () => {
-		snackbarService.show('Button clicked');
-	}
-
 	return (
 		<Fragment>
-
-			<Button
-				block={settingValues.block}
-				color={settingValues.color}
+			<IconButton
 				className={settingValues.className}
-				disabled={settingValues.disabled}
+				color={settingValues.color}
 				isActive={settingValues.isActive}
-				isRounded={settingValues.isRounded}
+				disabled={settingValues.disabled}
+				icon={homeSolidSvg}
 				variant={settingValues.variant}
-				onClick={handleClick}
-			>
-				some button text
-			</Button>
-
+			/>
 		</Fragment>
 	);
 }
 
-export const ShowcaseButton = withOptions<IButtonProps>(ShowcaseButtonBase, {
+export const IconButtonPage = withOptions<IIconButtonProps>(IconButtonPageBase, {
 	color: COLOR.primary,
 	variant: VARIANT.normal
 });