+++ src/showcase/index.ts
@@ -2,3 +2,4 @@ export * from './AboutPage';
 export * from './HomePage';
 export * from './GettingStartedPage';
 export * from './examples';
+export * from './showcase.models';