+++ src/showcase/showcase.models.ts
@@ -15,3 +15,15 @@ export class MenuModel implements ISidebarItem {
 	path: string;
 	items: Array<MenuModel>;
 }
+
+export class RouteModel {
+	constructor(dto: IMenuItem, parentPath?: string) {
+		this.path = [parentPath, dto.path !== undefined ? dto.path : dto.id].join("/");
+		this.componentKey = `${dto.id}Page`;
+		this.routes = dto.items && dto.items.map(item => new RouteModel(item, this.path));
+	}
+
+	path: string;
+	componentKey: string;
+	routes?: Array<RouteModel>;
+}